+++ CHANGES.md
@@ -5,8 +5,11 @@ CHANGES
 1.1-SNAPSHOT
 ------------
 
+- Added getURI method to WebSocket.
 - Added onStateChanged method to WebSocketListener.
 
 
 1.0 (2015-04-22)
 ----------------
+
+- The first release.